moved TestAddEdgesIfDoNotExist to db
